move eq to new file
