+++ src/operators/comparison/index.ts
@@ -9,18 +9,6 @@ export function between(column: string, x: string | number, y: string | number):
 	return `${column} BETWEEN '${x}' AND '${y}'`;
 }
 
-export class EqualsComparisonOperator {
-	public static eq(column: string, value: string | number): EqualsComparisonOperator {
-		return new EqualsComparisonOperator(column, value);
-	}
-
-	constructor(public readonly column: string, public readonly value: string | number) {}
-
-	public resolve(): string {
-		return `${this.column} = ${this.value}`;
-	}
-}
-
 export function exists(subquery: string | QueryBuilder): string {
 	if (subquery instanceof QueryBuilder) {
 		subquery = subquery.toSQL({ semicolon: false });
